+++ main.py
@@ -20,6 +20,19 @@ from helpers.recommender_utils import any_list_element_in_string
 from helpers.recommender_utils import get_cosine_score
 
 #-----------------------------------
+# Variables
+#-----------------------------------
+GEMINI_MODEL="gemini-2.0-flash-001"
+EMBEDDING_MODEL="text-embedding-005"
+COSINE_SCORE_THRESHOLD=0.65
+TEMPERATURE=0.1
+TOP_P=0.8
+TOP_K=25
+#STYLIST_PROMPT="Describe the clothes in the photo.  Description should only include clothing 'type', 'color', 'style', and 'design'. Make sure to only describe each individual article of clothing, and give a separate response."
+STYLIST_PROMPT="Can you describe the clothes in the photo, including style, color, and any designs?  Make sure to only describe each individual article of clothing, and give a separate response."
+
+
+#-----------------------------------
 # Initialize Vertex AI & Gemini
 #-----------------------------------
 PROJECT_ID = os.environ.get('MY_PROJECT_ID')  # @param {type:"string"}
@@ -38,7 +51,7 @@ vertexai.init(project=PROJECT_ID, location=LOCATION)
 
 #multimodal_model = GenerativeModel("gemini-1.0-pro-vision")
 multimodal_model = GenerativeModel(
-        "gemini-2.0-flash-001",
+        GEMINI_MODEL,
         system_instruction=[
             "You are a fashion stylist.",
             "Your mission is to describe the clothing you see.",
@@ -114,11 +127,10 @@ def get_text_embedding_from_text_embedding_model(
     """
     client = genai.Client()
     response = client.models.embed_content(
-        #model="text-embedding-005",
-        model="text-embedding-large-exp-03-07",
+        model=EMBEDDING_MODEL,
         contents=text,
         config=EmbedContentConfig(
-            task_type="RETRIEVAL_DOCUMENT",  # Optional
+            task_type="RETRIEVAL_QUERY",  # Optional
             output_dimensionality=768,  # Optional
         ),
     )
@@ -229,9 +241,9 @@ def get_user_query_text_embeddings(user_query: str) -> np.ndarray:
 def get_reference_image_description(image_filename: str) -> list:
   # Use a more deterministic configuration with a low temperature
   generation_config = GenerationConfig(
-    temperature=0.0,
-    top_p=0.8,
-    top_k=20,
+    temperature=TEMPERATURE,
+    top_p=TOP_P,
+    top_k=TOP_K,
     candidate_count=1,	# reponse
     max_output_tokens=512,
   )
@@ -323,7 +335,7 @@ def upload() -> str:
   jacket_word_list=[' jacket', ' coat', ' parka', ' blazer', ' vest']
   sweater_word_list=[' sweater', ' hoodie']
   shirt_word_list=[' t-shirt', ' shirt', ' tank top']
-  pant_word_list=[' pants', ' jeans', ' sweatpants', ' shorts', ' chinos', ' khakis']
+  pant_word_list=[' pants', ' jeans', ' sweatpants', ' shorts', ' chinos', ' khakis', 'trousers']
   shoe_word_list=[' shoes', ' sneakers', ' loafers', ' clogs']
   clothing_list=[hat_word_list, jacket_word_list, sweater_word_list, shirt_word_list, pant_word_list, shoe_word_list]