+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.6.1] - 2025-04-28
+### Changed
+- Variablize model config settings
+- Tuned Gemini model `GenerationConfig` settings & prompt
+- Embedding model `task_type` changed from `RETRIEVAL_DOCUMENT` to `RETRIEVAL_QUERY`
+
 ## [0.6.0] - 2025-04-26
 ### Changed
 - Now using `genai` instead of `vertexai.language_models` for embedding